tests/wpt/web-platform-tests/navigation-api/navigation-methods/return-value/forward-intercept-rejected.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/helpers.js"></script>

<script>
promise_test(async t => {
  // Wait for after the load event so that the navigation doesn't get converted
  // into a replace navigation.
  await new Promise(resolve => window.onload = () => t.step_timeout(resolve, 0));

  location.href = "#1";
  await navigation.back().committed;

  assert_equals(navigation.entries().length, 2);
  const [entry0, entry1] = navigation.entries();
  assert_equals((new URL(entry0.url)).hash, "");
  assert_equals((new URL(entry1.url)).hash, "#1");

  const err = new Error("boo!");
  const promise = Promise.reject(err);
  promise.catch(() => {}); // prevent unhandled rejection testharness.js errors
  navigation.onnavigate = e => e.intercept({ handler: () => promise });

  const result = navigation.forward();
  await assertCommittedFulfillsFinishedRejectsExactly(t, result, entry1, err);
  assert_equals(navigation.currentEntry, entry1);
}, "forward() promise rejection with rejected intercept()");
</script>
